refactor: rename number.ts to formatNumber.ts
