+++ packages/twenty-front/src/modules/ui/field/display/components/NumberDisplay.tsx
@@ -1,4 +1,4 @@
-import { formatNumber } from '~/utils/format/number';
+import { formatNumber } from '~/utils/format/formatNumber';
 
 import { EllipsisDisplay } from './EllipsisDisplay';
