+++ packages/twenty-front/src/utils/format/formatNumber.ts
